refactor: extract time formatter utility
